% Generated by floodflow doc builder: edit the roxygen in R/, not here.
\name{is_flood_project}
\alias{is_flood_project}
\title{Test whether an object is a flood project}
\description{
Test whether an object is a flood project
}
\arguments{
\item{x}{An object to test.}
}
\value{
A single logical value: \code{TRUE} if \code{x} inherits from class \code{flood_project}, otherwise \code{FALSE}.
}
\examples{
is_flood_project(flood_project())
is_flood_project(list())

}
\author{George Owusu}
