% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_project.R
\name{is_flood_project}
\alias{is_flood_project}
\title{Test whether an object is a flood project}
\usage{
is_flood_project(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A single logical value: \code{TRUE} if \code{x} inherits from class
  \code{flood_project}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a flood project
}
\examples{
is_flood_project(flood_project())
is_flood_project(list())

}
