% Generated by floodflow doc builder: edit the roxygen in R/, not here.
\name{print.flood_project}
\alias{print.flood_project}
\title{Print a flood project}
\description{
Print a flood project
}
\arguments{
\item{x}{A \code{flood_project} object.}
\item{...}{Ignored, present for S3 method consistency.}
}
\value{
The \code{flood_project} object \code{x}, returned invisibly. Called for the side effect of printing a compact summary to the console.
}
\examples{
print(flood_project(name = "Odaw basin"))

}
\author{George Owusu}
