% Generated by floodflow doc builder: edit the roxygen in R/, not here.
\name{tc_kirpich}
\alias{tc_kirpich}
\title{Time of concentration by the Kirpich method}
\description{
The Kirpich (1940) empirical formula for the time of concentration of a channelised catchment, in the SI form with length in metres. Time of concentration is the time for water to travel from the hydraulically most distant point to the outlet, a key control on peak discharge timing.
}
\arguments{
\item{length_m}{Flow-path length in metres.}
\item{slope}{Channel slope (dimensionless, m/m).}
}
\value{
Time of concentration in minutes.
}
\examples{
tc_kirpich(1500, 0.05)
}
\references{
Kirpich, Z. P. (1940) Time of concentration of small
  agricultural watersheds. Civil Engineering 10(6), 362.
}
\author{George Owusu}
